+++ src/main/java/id/co/indivara/jdt12/warehouseproject/repository/SupplyToWarehouseRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import javax.transaction.Transactional;
 
 @Transactional
-public interface SupplyToWarehouseRepository extends JpaRepository<SupplyToWarehouse, Integer> {
+public interface SupplyToWarehouseRepository extends JpaRepository<SupplyToWarehouse, Long> {
 }
